Updated modules (#83)
* fix: Prevent service data saving if api response failed

* #78 Change token display to a confirm dialog

* Revised font loading

* #74 feat: Added LeftClickAction to reminders

* Fixed module validations

* feat: Added reminder stack direction

* fix: Updated discord username

* feat: Added color & opacity options for top timeline

* chore: bring SCT up to latest view/control standards

* feat: Give views the possibility to change label width

* fix: Fontsize as float

* fix: fixed possible NPE

* bump versions

+++ Estreya.BlishHUD.Shared/Utils/FontUtils.cs
@@ -3,12 +3,17 @@
     using Blish_HUD;
     using Microsoft.Xna.Framework;
     using Microsoft.Xna.Framework.Graphics;
+    using MonoGame.Extended.BitmapFonts;
+    using MonoGame.Extended.TextureAtlases;
     using SpriteFontPlus;
+    using System;
+    using System.Collections.Generic;
     using System.IO;
+    using System.Linq;
 
     public static class FontUtils
     {
-        public static SpriteFont FromTrueTypeFont(byte[] ttfData, int fontSize, int bitmapWidth, int bitmapHeight)
+        public static SpriteFont FromTrueTypeFont(byte[] ttfData, float fontSize, int bitmapWidth, int bitmapHeight)
         {
             TtfFontBakerResult fontBakeResult = TtfFontBaker.Bake(ttfData,
                 fontSize,
